Capitalized first letter of folders in src/ and moved page components to src/Pages
