+++ e-commerce-app/src/Pages/ProductPage.js
@@ -1,8 +1,8 @@
 import { useParams } from "react-router-dom";
 import { useContext, useEffect, useState } from "react";
 import { getProductInfo } from "./apis/product_api";
-import Spinner from "./components/Spinner";
-import CartContext from "./context/CartContext";
+import Spinner from "../Components/Spinner";
+import CartContext from "../Context/CartContext";
 
 import "./Styles/ProductPage.css";
 import "./Styles/common.css";
